fix(urgentie): address PR review findings
- Fix DateTime/DateTimeOffset mismatch in UrgentieBerekenService
  (construct endOfDay as DateTimeOffset consistently)
- Remove unused BadgeStatus import in UrgentieBadge.vue
- Move UrgentieInfo and MyInterneTaakOverviewItem types to src/types/internetaken.ts
  (decouple service layer from Vue SFC exports)
- Use Math.ceil instead of Math.round for days conversion in UrgentieBadge
  (prevents '0d' edge case, consistent with backend ceiling)

Task: #498
